requirements.txt
django==4.0.5
selenium==4.2.0
fabric2==2.7.0
